Add support for one way flights
Southwest provides and displays one-way flights, so should we. This
commit adds support for the `--one-way` flag which changes the form
submission to only include single flights between two airports.

The flag removes conflicting options added by other flags such as
`--return-date` in order to prevent errors and incorrect form
submissions.

This feature addresses #23 and produces valid data (could use some
testing).

Also in the commit is a documentation (comments) for #24 which notes
that there is an implicit use of a blessed widget method through
contrib inheriting blessed > List.

Closes #25.