Add support for outbound/returen date validation
This commit adds two functions for checking the valid starting - ending
dates for Southwest flights:

`checkIfPastDates`: Checks if either outbound or return are before the
durrent date (year, month, day - not hour)

`checkIfFutureDates`: First finds `#lastBookableDate` from the Southwest
website, then uses it to check if either outbound or return dates are
after the last bookable flight.

Checking for past dates is done immediately after command line args are
parsed. If they are invalid, and error is thrown.
Note: this could be handled better with a friendlier message to the
user.

Checking for future dates is done after the `#lastBookableDate` element
is found on southwest.com.
If there are invalid dates, it displays in the Log.
Note: The setInterval created by the waypoints casuses the message to be
added to the Log multiple times due to screen.render() being called.

Would appreciate any feedback!